Copied lambda-ccc/test/fft_test.hs to fft-ccc.hs.
I decided to pull this file up to the top level of the TreeViz project,
since it really is intended just as a test vehicle for different
expressions of FFT, and not necessarily as compilable, via Conal's
machinery.

I've stated my intent, in the header of the lambda-ccc/test/fft_test.hs
file, asking that further general explorations in FFT expression be done
in fft-ccc.hs, instead.